[IDLE-278] feat: 업데이트 알림 홈화면 적용 및 컴포넌트 공통화 - firebase 통신을 바탕으로 홈 화면에 업데이트 alert 적용 - 확안 또는 확인/취소 AlertDialog 를 common_utils 파일에 공통화하여 모두 적용

+++ app-ui/lib/ui/setting/setting_screen.dart
@@ -8,6 +8,7 @@ import 'package:mybrary/ui/common/layout/default_layout.dart';
 import 'package:mybrary/ui/profile/profile_edit/profile_edit_screen.dart';
 import 'package:mybrary/ui/setting/components/account_withdrawal.dart';
 import 'package:mybrary/ui/setting/components/login_info.dart';
+import 'package:mybrary/utils/logics/common_utils.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class SettingScreen extends StatelessWidget {
@@ -216,14 +217,14 @@ class SettingScreen extends StatelessWidget {
           actions: [
             Row(
               children: [
-                _confirmButton(
+                confirmButton(
                   onTap: () {
                     Navigator.of(context).pop();
                   },
                   buttonText: '취소',
                   isCancel: true,
                 ),
-                _confirmButton(
+                confirmButton(
                   onTap: () async {
                     await secureStorage.deleteAll();
 
@@ -236,6 +237,8 @@ class SettingScreen extends StatelessWidget {
                   },
                   buttonText: '로그아웃',
                   isCancel: false,
+                  confirmButtonColor: commonRedColor,
+                  confirmButtonText: commonWhiteColor,
                 ),
               ],
             ),
@@ -268,37 +271,6 @@ class SettingScreen extends StatelessWidget {
               'assets/svg/icon/profile_menu_arrow.svg',
             ),
           ],
-        ),
-      ),
-    );
-  }
-
-  Widget _confirmButton({
-    required GestureTapCallback? onTap,
-    required String buttonText,
-    required bool isCancel,
-  }) {
-    return Expanded(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 4.0),
-        child: InkWell(
-          onTap: onTap,
-          child: Container(
-            height: 46.0,
-            decoration: BoxDecoration(
-              color: isCancel ? greyF1F2F5 : commonRedColor,
-              borderRadius: BorderRadius.circular(4.0),
-            ),
-            child: Center(
-              child: Text(
-                buttonText,
-                style: commonSubBoldStyle.copyWith(
-                  color: isCancel ? commonBlackColor : commonWhiteColor,
-                  fontSize: 14.0,
-                ),
-              ),
-            ),
-          ),
         ),
       ),
     );
